requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
google-genai>=0.3.0
Werkzeug==3.0.1
supabase==2.27.2

